Move validator calls back to UserProfileService for AB#16696.
