+++ Apps/GatewayApi/src/Controllers/UserProfileController.cs
@@ -44,7 +44,6 @@ namespace HealthGateway.GatewayApi.Controllers
         private readonly IHttpContextAccessor httpContextAccessor;
         private readonly IUserEmailService userEmailService;
         private readonly IUserProfileService userProfileService;
-        private readonly IUserProfileValidatorService userProfileValidatorService;
         private readonly IUserSmsService userSmsService;
         private readonly IUserPreferenceService userPreferenceService;
         private readonly ILegalAgreementService legalAgreementService;
@@ -53,17 +52,14 @@ namespace HealthGateway.GatewayApi.Controllers
         /// Initializes a new instance of the <see cref="UserProfileController"/> class.
         /// </summary>
         /// <param name="userProfileService">The injected user profile service.</param>
-        /// <param name="userProfileValidatorService">The injected user profile validator service.</param>
         /// <param name="httpContextAccessor">The injected http context accessor provider.</param>
         /// <param name="userEmailService">The injected user email service.</param>
         /// <param name="userSmsService">The injected user sms service.</param>
         /// <param name="authenticationDelegate">The injected authentication delegate.</param>
         /// <param name="userPreferenceService">The injected user preference service.</param>
         /// <param name="legalAgreementService">The injected legal agreement service.</param>
-#pragma warning disable S107 // The number of DI parameters should be ignored
         public UserProfileController(
             IUserProfileService userProfileService,
-            IUserProfileValidatorService userProfileValidatorService,
             IHttpContextAccessor httpContextAccessor,
             IUserEmailService userEmailService,
             IUserSmsService userSmsService,
@@ -72,7 +68,6 @@ namespace HealthGateway.GatewayApi.Controllers
             ILegalAgreementService legalAgreementService)
         {
             this.userProfileService = userProfileService;
-            this.userProfileValidatorService = userProfileValidatorService;
             this.httpContextAccessor = httpContextAccessor;
             this.userEmailService = userEmailService;
             this.userSmsService = userSmsService;
@@ -158,7 +153,7 @@ namespace HealthGateway.GatewayApi.Controllers
         [Authorize(Policy = UserProfilePolicy.Read)]
         public async Task<RequestResult<bool>> Validate(string hdid, CancellationToken ct)
         {
-            return await this.userProfileValidatorService.ValidateMinimumAgeAsync(hdid, ct);
+            return await this.userProfileService.ValidateMinimumAgeAsync(hdid, ct);
         }
 
         /// <summary>
@@ -425,7 +420,7 @@ namespace HealthGateway.GatewayApi.Controllers
         [Authorize]
         public async Task<ActionResult<bool>> IsValidPhoneNumber(string phoneNumber, CancellationToken ct)
         {
-            return await this.userProfileValidatorService.IsPhoneNumberValidAsync(phoneNumber, ct);
+            return await this.userProfileService.IsPhoneNumberValidAsync(phoneNumber, ct);
         }
     }
 }
